requirements.txt
fastapi
uvicorn
openai
python-dotenv
PyPDF2
python-docx
pydantic
aiofiles
requests
logging
datetime
typing
python-multipart